fix(wextra): suppress -Wextra warnings in reduction and sliding_window [r45]
- generic_reductions: use int rank (cast from size_t) to fix sign-compare
  in all rank comparisons; cast input_shape.size() in standalone loops
- reduction_common: same int rank pattern; cast dim.size() in loop bound
- argmax: ideal_red_dim_units: int → uint32_t to match red_dim_units type
- accumulation_common: cast FOUR_DIMENSIONS to int32_t in comparisons
- accumulation_program_factory: cast input_shape.rank() to int32_t in loop
- prod_nc_device_operation: int → size_t loop over input_shape.rank()
- reduce_op_multi_core_h/w, welford: int y/x → auto for CoreCoord loops
- halo_device_operation: add missing .output_memory_config = {} to fix
  -Wmissing-designated-field-initializers
- untilize_with_halo_program_factory: int → size_t for cores.size() loop
- sliding_window: int → size_t for vector size loops;
  uint32_t i → int i in replicate_config (factor is int)

+++ ttnn/cpp/ttnn/operations/reduction/accumulation/accumulation_common.cpp
@@ -27,7 +27,7 @@ Tensor preprocess_input_tensor(
         int32_t final_cum_axis = cum_axis;
         if (input_rank < FOUR_DIMENSIONS) {
             ttnn::SmallVector<uint32_t> new_dims = {};
-            for (int32_t i = input_rank; i < FOUR_DIMENSIONS; ++i) {
+            for (int32_t i = static_cast<int32_t>(input_rank); i < static_cast<int32_t>(FOUR_DIMENSIONS); ++i) {
                 new_dims.push_back(1);
             }
             new_dims.insert(new_dims.end(), input_shape.cbegin(), input_shape.cend());
@@ -62,9 +62,9 @@ Tensor postprocess_output_tensor(
     const int32_t& original_rank) {
     Tensor processed_tensor = output_tensor;
 
-    if (original_rank - dim < FOUR_DIMENSIONS) {
+    if (original_rank - dim < static_cast<int32_t>(FOUR_DIMENSIONS)) {
         processed_tensor = ttnn::permute(processed_tensor, permutation, processed_tensor.memory_config());
-        if (original_rank < FOUR_DIMENSIONS) {
+        if (original_rank < static_cast<int32_t>(FOUR_DIMENSIONS)) {
             processed_tensor = ttnn::reshape(processed_tensor, original_shape);
         }
     }
